+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     - added new Laser Temperature panel within Laser Control widget
     - added Laser Temperature to StatusBar
     - added EDC for IMX385
+    - LaserTemperatureFeature to ctl
+        - added XS Laser TEC Mode
 - 2024-03-13 4.0.50
     - fixed changelog (2024)
     - SRM / Wavenumber button tweaks (also location)